requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
email-validator==2.1.0.post1
flask==3.0.2
Flask-Cors==4.0.0
Flask-DebugToolbar==0.14.1
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
psycopg2-binary==2.9.9
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
sqlalchemyseed==2.0.0
typing-extensions==4.10.0
werkzeug==3.0.1
wtforms==3.1.2
zipp==3.18.1
